update additionl harvest costs script
Updated this script to add CTL lop and scatter values to match CEC report (though there were no CTL lop and scatter prescriptions for the project).

The script was also updated to back up any tables it replaces, and instead of replacing with a new table, it uses a join/update query to update values in the tables. This was in order to hopefully avoid issues with the primary key values not remaining in the scenario_additional_harvest_costs table when a table is added via R, which may have caused core analysis to hang during audit.  The primary key values should now remain.